+++ src/main/java/com/lesniewicz/api/entity/Payment.java
@@ -1,9 +1,9 @@
 package com.lesniewicz.api.entity;
 
-import javax.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.persistence.*;
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
@@ -16,7 +16,7 @@ public class Payment {
     @Id
     @Column(nullable = false, updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer paymentId;
+    private Long paymentId;
 
     @Column(nullable = false, precision = 7, scale = 2)
     private BigDecimal amount;
