+++ src/main/java/com/lesniewicz/api/repository/InventoryRepository.java
@@ -3,5 +3,5 @@ package com.lesniewicz.api.repository;
 import com.lesniewicz.api.entity.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+public interface InventoryRepository extends JpaRepository<Inventory, Long> {
 }
