+++ src/main/java/com/lesniewicz/api/repository/LanguageRepository.java
@@ -5,5 +5,5 @@ import com.lesniewicz.api.entity.Language;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface LanguageRepository extends JpaRepository<Language, Integer> {
+public interface LanguageRepository extends JpaRepository<Language, Long> {
 }
